refactor(buildsource): split .abicheck.yml config schema out of inline.py (G38 Phase 15 prereq)
docs/contribute/plans/g38-bundle-facts-model-and-multibuild-comparability.md's
Phase 15 needs new .abicheck.yml (bundle_variants:) parsing surface, but
every candidate host file was at or within a couple lines of the
AI-readiness 2000-line hard cap. inline.py was the cleanest split point:
its BuildConfig schema (parsing/validating a raw .abicheck.yml mapping)
shares no state with the file's own inline build/source collection
pipeline (collect_inline_pack and everything it calls) -- confirmed by a
full-file grep before the split. Moved BuildConfig/load_build_config/
discover_build_config into a new sibling module, build_config.py, the
same mechanical-extraction shape build_config_schema.py/build_config_io.py
were already split out for. inline.py re-exports all three for
back-compat (from abicheck.buildsource.inline import BuildConfig keeps
working); new code should import from build_config.py directly.

inline.py: 1975 -> 1414 lines, well clear of the cap. This is only the
file-split prerequisite -- Phase 15's actual bundle_variants: feature and
CLI dispatch are separate, not-yet-attempted work.
