+++ scripts/CLAUDE.md
@@ -47,7 +47,7 @@ Each must run with Python 3.10+ and the package installed in dev mode
 | `gen_action_reference.py` | Regenerates the exhaustive GitHub Action inputs/outputs reference (`docs/reference/github-action-inputs.md`) from `action.yml`'s `description`/`required`/`default` fields — the single source of truth `docs/use/github-action.md`'s narrative previously restated by hand. `--check` gates sync; mirrored in `tests/test_action_reference.py`. Run after adding, renaming, or changing the default of an Action input/output. | manual |
 | `gen_cli_reference.py` | Regenerates the exhaustive CLI command/option reference (`docs/reference/cli-reference.md`) from the live Click command tree (`abicheck/cli.py` + registered sibling modules) — every command/subcommand/argument/option with its exact `help=` text, default, and choices. `--check` gates sync; mirrored in `tests/test_cli_reference.py`. Run after adding/renaming a command or option, or changing its `help=`/default. | manual |
 | `gen_python_api_reference.py` | Regenerates the exhaustive Python API reference (`docs/reference/python-api-reference.md`) from every name in `abicheck.service.__all__` — full function signatures and dataclass field lists. `--check` gates sync; mirrored in `tests/test_python_api_reference.py`. Run after changing `service.__all__` or a public function/dataclass's signature. | manual |
-| `gen_config_reference.py` | Regenerates the exhaustive `.abicheck.yml` key/type reference (`docs/reference/config-keys-reference.md`) from `abicheck.buildsource.inline.BuildConfig`'s strict-schema registries (`_KNOWN_TOP_KEYS`/`_KNOWN_BLOCK_KEYS`/`_BOOL_SUBKEYS`/`_STR_SUBKEYS`/`_LIST_SUBKEYS`). `--check` gates sync; mirrored in `tests/test_config_reference.py`. Run after adding/renaming a config key or changing its required type. | manual |
+| `gen_config_reference.py` | Regenerates the exhaustive `.abicheck.yml` key/type reference (`docs/reference/config-keys-reference.md`) from `abicheck.buildsource.build_config.BuildConfig`'s strict-schema registries (`_KNOWN_TOP_KEYS`/`_KNOWN_BLOCK_KEYS`/`_BOOL_SUBKEYS`/`_STR_SUBKEYS`/`_LIST_SUBKEYS`). `--check` gates sync; mirrored in `tests/test_config_reference.py`. Run after adding/renaming a config key or changing its required type. | manual |
 | `gen_agent_skills.py` | Publishes `skills-src/` as self-contained Agent Skills (ADR-058 / G36 P0.3) into the three generated trees `.agents/skills/`, `.claude/skills/`, and `.gemini/skills/` — no longer committed (2026-08-21 ADR-058 amendment); build output, regenerated on demand. Resolves each skill's cited `shared/*.md` Layer-B fragments **transitively** (a fragment may cite a fragment), **copies** them into that skill's own `references/shared/` (never symlinks — a marketplace zip, a `skills.sh` subset install, and a Windows checkout each break a live cross-skill link differently), rewrites every internal link to its generated position, and rewrites `docs/` links to the published site using `mkdocs.yml`'s own `site_url`/`use_directory_urls` rather than a hardcoded URL. Fails on a missing fragment reference, an orphaned (never-cited) fragment, and a link escaping the skill's own installed directory. `--check` now validates `skills-src/`'s own internal consistency (a real render succeeds, is deterministic, and is byte-identical across the three target roots it would produce) rather than diffing against on-disk output that no longer exists; mirrored in `tests/test_gen_agent_skills.py` and wired as `verify.py`'s `agent-skills-generated` step. | `verify.py --only agent-skills-generated` (CI: the `ai-readiness` job's second gate step, alongside `docs-contract`/`repo-facts`); manual after any `skills-src/` edit. |
 | `install_dev_skill.py` | Thin CLI wrapper around `gen_agent_skills.py`'s `render_all`/`write_trees` for local dev use: materializes the generated skill tree(s) on disk (`.agents/skills/`, `.claude/skills/`, `.gemini/skills/` — all gitignored) so a contributor can point Claude Code/Codex/Cursor/Gemini CLI, or a manual Harbor task run, at an installed skill. `--target {codex,claude,gemini,all}` (`codex` selects the portable `.agents/skills/` tree Copilot/Cursor also read). No rendering logic of its own. | manual, after a `skills-src/` edit, before exercising an installed skill locally. |
 | `gen_platform_matrix.py` | Regenerates the host-OS × binary-format capability matrix — `docs/reference/platforms.md`'s "Quick Reference: What Works Where" section, spliced between `<!-- BEGIN/END GENERATED: platform-matrix -->` sentinels — from `platform_capabilities.py`. `--check` gates sync; mirrored in `tests/test_platform_matrix.py`. Run after a symbol/type-diff capability or tooling requirement changes for a host/format pair (ADR-051 Stage 2). | manual |
